AP/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
python-dotenv